fo/requirements.txt
Django>=1.3
MySQL-python==1.2.3
django-countries
django-coverage
django-reversion
django-tastypie
mimeparse>=0.1.3
python-dateutil==1.5
python-digest>=1.7
wsgiref>=0.1.2
python-openid>=2.2
django-social-auth
oauth2>=1.5.167
